refactor: move unwrapToolResponse into AnthropicClient.ts
Per reviewer feedback, inline the utility into AnthropicClient.ts
instead of keeping it as a standalone file. Update imports in
aisdk.ts and the unit test accordingly.
